Create SECURITY.md
Adding a SECURITY.md file to the Space Warp project. This file provides guidelines for reporting security vulnerabilities in the project, as well as our disclosure policy and contact information (when we get some contact info).

By having a SECURITY.md file, we can ensure that our users and contributors are aware of our security policies and know how to report vulnerabilities in a responsible way. This helps us to protect our users and maintain the security of the project.

The SECURITY.md file that I have created includes information on reporting vulnerabilities, and contact information for security concerns. It also includes a "Supported Versions" section, which explains that we only support the latest released version of the mod.