+ console is remade in uitk + optimized event system
mostly just rewrote it, the message event system is pretty optimized. it unloads when the window is closed, and stuff. is the code bad: yes. does it work really well: yes.

only currently known bug is the search bar. there are drag issues when you click on the textinput field where it gets stuck to your mouse.