Rework assembly loading process when running in the editor.
This allows assemblyies complied and/or loaded by Unity during domain reload to
be used directly.  These can be precompiled DLLs under Assets, or
assemblies built with `asmdef`.

Unity will looad an dll file under Assets by default.  So the strategy
is to just assume they are loaded and reflect through loaded assmeblies
to find the corresponding MainAssembly.

This still requires swinfo.json to be placed under Assets/Mods to
trigger redux to look for it, but probably should't require a built
asset catalog.  Addressables in the editor should be able to find the
assets and load them through normal editor asset loading process.